fix some things
