+++ setup.py
@@ -8,7 +8,7 @@ def read_long_description(file_path):
 
 setup(
     name="scrapy-puppeteer-client",
-    version="0.3.1",
+    version="0.3.3",
     description="A library to use Puppeteer-managed browser in Scrapy spiders",
     long_description=read_long_description("README.md"),
     long_description_content_type="text/markdown",
